Use idempotency id for session id
By using the idempotency id the usage of the library can be even simpler. Another improvement that can be added later on is to move the Idempotency id to a field of the response. Then the DataBundleSender only needs a dependency on the response and can use the value of Idempotency Id field for the session value.